BJ-902: Move thrift converters to the new package
