+++ src/main/java/com/rbkmoney/threeds/server/config/CacheServiceConfig.java
@@ -2,8 +2,8 @@ package com.rbkmoney.threeds.server.config;
 
 import com.rbkmoney.damsel.three_ds_server_storage.CardRangesStorageSrv;
 import com.rbkmoney.damsel.three_ds_server_storage.RReqTransactionInfoStorageSrv;
-import com.rbkmoney.threeds.server.converter.CardRangesConverter;
-import com.rbkmoney.threeds.server.converter.RReqTransactionInfoConverter;
+import com.rbkmoney.threeds.server.converter.thrift.CardRangesConverter;
+import com.rbkmoney.threeds.server.converter.thrift.RReqTransactionInfoConverter;
 import com.rbkmoney.threeds.server.service.cache.CacheService;
 import com.rbkmoney.threeds.server.service.cache.InMemoryCacheService;
 import com.rbkmoney.threeds.server.service.cache.ThreeDsServerStorageCacheService;
