+++ src/main/java/com/rbkmoney/threeds/server/converter/thrift/RReqTransactionInfoConverter.java
@@ -1,4 +1,4 @@
-package com.rbkmoney.threeds.server.converter;
+package com.rbkmoney.threeds.server.converter.thrift;
 
 import com.rbkmoney.threeds.server.domain.acs.AcsDecConInd;
 import com.rbkmoney.threeds.server.domain.device.DeviceChannel;
