+++ src/main/java/com/rbkmoney/threeds/server/service/cache/ThreeDsServerStorageCacheService.java
@@ -6,8 +6,8 @@ import com.rbkmoney.damsel.three_ds_server_storage.CardRangesStorageSrv;
 import com.rbkmoney.damsel.three_ds_server_storage.GetCardRangesRequest;
 import com.rbkmoney.damsel.three_ds_server_storage.GetCardRangesResponse;
 import com.rbkmoney.damsel.three_ds_server_storage.RReqTransactionInfoStorageSrv;
-import com.rbkmoney.threeds.server.converter.CardRangesConverter;
-import com.rbkmoney.threeds.server.converter.RReqTransactionInfoConverter;
+import com.rbkmoney.threeds.server.converter.thrift.CardRangesConverter;
+import com.rbkmoney.threeds.server.converter.thrift.RReqTransactionInfoConverter;
 import com.rbkmoney.threeds.server.domain.CardRange;
 import com.rbkmoney.threeds.server.dto.RReqTransactionInfo;
 import com.rbkmoney.threeds.server.exeption.ThreeDsServerStorageException;
