+++ src/test/java/com/rbkmoney/threeds/server/service/ThreeDsServerStorageCacheServiceTest.java
@@ -3,8 +3,8 @@ package com.rbkmoney.threeds.server.service;
 import com.rbkmoney.damsel.three_ds_server_storage.CardRangesStorageSrv;
 import com.rbkmoney.damsel.three_ds_server_storage.GetCardRangesResponse;
 import com.rbkmoney.damsel.three_ds_server_storage.RReqTransactionInfoStorageSrv;
-import com.rbkmoney.threeds.server.converter.CardRangesConverter;
-import com.rbkmoney.threeds.server.converter.RReqTransactionInfoConverter;
+import com.rbkmoney.threeds.server.converter.thrift.CardRangesConverter;
+import com.rbkmoney.threeds.server.converter.thrift.RReqTransactionInfoConverter;
 import com.rbkmoney.threeds.server.domain.acs.AcsDecConInd;
 import com.rbkmoney.threeds.server.domain.device.DeviceChannel;
 import com.rbkmoney.threeds.server.dto.RReqTransactionInfo;
